frontend: address PR feedback
